several fixes for the subnav, scrollspy, + handling of hash links
lots of changes to deal with quirks arising with having scrollspy,
subnav, and custom hash link handling all going at once.

One big problem with scrollspy was that the twitter bootstrap version
always assumed that scrollspy would be initialed with scrollTop == 0,
and if it wasn't, then the stored offsets of elements were all wrong.
I've added a patched version of scrollspy to resolve that issue.